+++ src/core/elements/utils.cpp
@@ -57,7 +57,7 @@ Distances compute_distances_from_state(const AdjList& adj_list, int source) {
 
 PairwiseDistances compute_pairwise_distances(const AdjList& adj_list) {
     PairwiseDistances pairwise_distances;
-    for (int source = 0; source < adj_list.size(); ++source) {
+    for (int source = 0; source < static_cast<int>(adj_list.size()); ++source) {
         pairwise_distances.push_back(compute_distances_from_state(adj_list, source));
     }
     return pairwise_distances;
